feat: export ensand account resolution into a hook
