+++ ui/pages/bridge/hooks/useDestinationAccountResolution.ts
@@ -1,0 +1,129 @@
+import { useMemo, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { isEthAddress } from '../../../../app/scripts/lib/multichain/address';
+import { DestinationAccount } from '../prepare/types';
+import {
+  getDomainResolutions,
+  lookupDomainName,
+  resetDomainResolution,
+  initializeDomainSlice,
+} from '../../../ducks/domains';
+import { isSolanaAddress } from '../../../../shared/lib/multichain/accounts';
+import { InternalAccount } from '@metamask/keyring-internal-api';
+
+interface UseDestinationAccountResolutionProps {
+  searchQuery: string;
+  isDestinationSolana: boolean;
+  accounts: InternalAccount[];
+}
+
+interface UseDestinationAccountResolutionResult {
+  isValidAddress: boolean;
+  isValidEnsName: boolean;
+  externalAccount: DestinationAccount | null;
+}
+
+export const useDestinationAccountResolution = ({
+  searchQuery,
+  isDestinationSolana,
+  accounts,
+}: UseDestinationAccountResolutionProps): UseDestinationAccountResolutionResult => {
+  const dispatch = useDispatch();
+  const domainResolutions = useSelector(getDomainResolutions) || [];
+
+  // Initialize domain slice on mount
+  useEffect(() => {
+    dispatch(initializeDomainSlice());
+  }, [dispatch]);
+
+  // Check if search query is a valid address
+  const isValidAddress = useMemo(() => {
+    const trimmedQuery = searchQuery.trim();
+    if (!trimmedQuery) {
+      return false;
+    }
+
+    return isDestinationSolana
+      ? isSolanaAddress(trimmedQuery)
+      : isEthAddress(trimmedQuery);
+  }, [searchQuery, isDestinationSolana]);
+
+  // Check if search query is a valid ENS name
+  const isValidEnsName = useMemo(() => {
+    if (isDestinationSolana) {
+      return false;
+    }
+    const trimmedQuery = searchQuery.trim();
+    if (!trimmedQuery) {
+      return false;
+    }
+    return trimmedQuery.endsWith('.eth');
+  }, [searchQuery, isDestinationSolana]);
+
+  // Lookup ENS name when we detect a valid ENS name
+  useEffect(() => {
+    if (isValidEnsName) {
+      dispatch(lookupDomainName(searchQuery.trim()));
+    } else {
+      dispatch(resetDomainResolution());
+    }
+  }, [dispatch, isValidEnsName, searchQuery]);
+
+  // Create an external account object if valid address is not in internal accounts
+  const externalAccount = useMemo(() => {
+    if (!isValidAddress && !isValidEnsName) {
+      return null;
+    }
+
+    // If it's a valid ENS name and we have resolutions, use the resolved address
+    if (isValidEnsName && domainResolutions.length > 0) {
+      const { resolvedAddress } = domainResolutions[0];
+      const ensName = searchQuery.trim();
+
+      const addressExists = accounts.some(
+        (account) =>
+          account.address.toLowerCase() === resolvedAddress.toLowerCase(),
+      );
+
+      if (addressExists) {
+        return null;
+      }
+
+      return {
+        address: resolvedAddress,
+        metadata: {
+          name: ensName,
+        },
+        isExternal: true,
+      };
+    }
+
+    // For regular addresses
+    if (isValidAddress) {
+      const address = searchQuery.trim();
+      const addressExists = accounts.some(
+        (account) => account.address.toLowerCase() === address.toLowerCase(),
+      );
+
+      if (addressExists) {
+        return null;
+      }
+
+      return {
+        address,
+        metadata: {
+          name: address,
+        },
+        isExternal: true,
+      };
+    }
+
+    return null;
+  }, [accounts, isValidAddress, isValidEnsName, searchQuery, domainResolutions]);
+
+  return {
+    isValidAddress,
+    isValidEnsName,
+    externalAccount,
+  };
+};
